[Security Solution][Endpoint][Admin] Refactor policy details protections (#94970)
