+++ x-pack/plugins/security_solution/public/management/pages/policy/view/policy_forms/protections/malware.tsx
@@ -5,332 +5,28 @@
  * 2.0.
  */
 
-import React, { useCallback, useMemo } from 'react';
-import { useDispatch } from 'react-redux';
+import React from 'react';
 import { i18n } from '@kbn/i18n';
 import { FormattedMessage } from '@kbn/i18n/react';
-import {
-  EuiCallOut,
-  EuiCheckbox,
-  EuiRadio,
-  EuiSpacer,
-  EuiSwitch,
-  EuiText,
-  EuiTextArea,
-  htmlIdGenerator,
-  EuiFlexGroup,
-  EuiFlexItem,
-  EuiIconTip,
-} from '@elastic/eui';
-import { cloneDeep } from 'lodash';
-import styled from 'styled-components';
+import { EuiCallOut, EuiSpacer } from '@elastic/eui';
 import { APP_ID } from '../../../../../../../common/constants';
 import { SecurityPageName } from '../../../../../../app/types';
-import {
-  Immutable,
-  OperatingSystem,
-  ProtectionModes,
-} from '../../../../../../../common/endpoint/types';
+import { Immutable, OperatingSystem } from '../../../../../../../common/endpoint/types';
 import { MalwareProtectionOSes, OS } from '../../../types';
-import { ConfigForm, ConfigFormHeading } from '../../components/config_form';
-import { policyConfig } from '../../../store/policy_details/selectors';
-import { usePolicyDetailsSelector } from '../../policy_hooks';
+import { ConfigForm } from '../../components/config_form';
 import { LinkToApp } from '../../../../../../common/components/endpoint/link_to_app';
 import { useLicense } from '../../../../../../common/hooks/use_license';
-import { AppAction } from '../../../../../../common/store/actions';
-import { SupportedVersionNotice } from './supported_version';
-
-export const RadioFlexGroup = styled(EuiFlexGroup)`
-  .no-right-margin-radio {
-    margin-right: 0;
-  }
-  .no-horizontal-margin-radio {
-    margin: ${(props) => props.theme.eui.ruleMargins.marginSmall} 0;
-  }
-`;
-
-const OSes: Immutable<MalwareProtectionOSes[]> = [OS.windows, OS.mac];
-const protection = 'malware';
-
-const ProtectionRadio = React.memo(
-  ({ protectionMode, label }: { protectionMode: ProtectionModes; label: string }) => {
-    const policyDetailsConfig = usePolicyDetailsSelector(policyConfig);
-    const dispatch = useDispatch<(action: AppAction) => void>();
-    const radioButtonId = useMemo(() => htmlIdGenerator()(), []);
-    // currently just taking windows.malware, but both windows.malware and mac.malware should be the same value
-    const selected = policyDetailsConfig && policyDetailsConfig.windows.malware.mode;
-    const isPlatinumPlus = useLicense().isPlatinumPlus();
-
-    const handleRadioChange = useCallback(() => {
-      if (policyDetailsConfig) {
-        const newPayload = cloneDeep(policyDetailsConfig);
-        for (const os of OSes) {
-          newPayload[os][protection].mode = protectionMode;
-          if (isPlatinumPlus) {
-            if (protectionMode === ProtectionModes.prevent) {
-              newPayload[os].popup[protection].enabled = true;
-            } else {
-              newPayload[os].popup[protection].enabled = false;
-            }
-          }
-        }
-        dispatch({
-          type: 'userChangedPolicyConfig',
-          payload: { policyConfig: newPayload },
-        });
-      }
-    }, [dispatch, protectionMode, policyDetailsConfig, isPlatinumPlus]);
-
-    /**
-     *  Passing an arbitrary id because EuiRadio
-     *  requires an id if label is passed
-     */
-
-    return (
-      <EuiRadio
-        className="policyDetailsProtectionRadio"
-        label={label}
-        id={radioButtonId}
-        checked={selected === protectionMode}
-        onChange={handleRadioChange}
-        disabled={selected === ProtectionModes.off}
-      />
-    );
-  }
-);
-
-ProtectionRadio.displayName = 'ProtectionRadio';
+import { RadioButtons } from '../components/radio_buttons';
+import { UserNotification } from '../components/user_notification';
+import { ProtectionSwitch } from '../components/protection_switch';
 
 /** The Malware Protections form for policy details
  *  which will configure for all relevant OSes.
  */
 export const MalwareProtections = React.memo(() => {
-  const policyDetailsConfig = usePolicyDetailsSelector(policyConfig);
-  const dispatch = useDispatch<(action: AppAction) => void>();
-  // currently just taking windows.malware, but both windows.malware and mac.malware should be the same value
-  const selected = policyDetailsConfig && policyDetailsConfig.windows.malware.mode;
-  const userNotificationSelected =
-    policyDetailsConfig && policyDetailsConfig.windows.popup.malware.enabled;
-  const userNotificationMessage =
-    policyDetailsConfig && policyDetailsConfig.windows.popup.malware.message;
+  const OSes: Immutable<MalwareProtectionOSes[]> = [OS.windows, OS.mac];
+  const protection = 'malware';
   const isPlatinumPlus = useLicense().isPlatinumPlus();
-
-  const radios: Immutable<
-    Array<{
-      id: ProtectionModes;
-      label: string;
-      protection: 'malware';
-    }>
-  > = useMemo(() => {
-    return [
-      {
-        id: ProtectionModes.detect,
-        label: i18n.translate('xpack.securitySolution.endpoint.policy.details.detect', {
-          defaultMessage: 'Detect',
-        }),
-        protection: 'malware',
-      },
-      {
-        id: ProtectionModes.prevent,
-        label: i18n.translate('xpack.securitySolution.endpoint.policy.details.prevent', {
-          defaultMessage: 'Prevent',
-        }),
-        protection: 'malware',
-      },
-    ];
-  }, []);
-
-  const handleSwitchChange = useCallback(
-    (event) => {
-      if (policyDetailsConfig) {
-        const newPayload = cloneDeep(policyDetailsConfig);
-        if (event.target.checked === false) {
-          for (const os of OSes) {
-            newPayload[os][protection].mode = ProtectionModes.off;
-            if (isPlatinumPlus) {
-              newPayload[os].popup[protection].enabled = event.target.checked;
-            }
-          }
-        } else {
-          for (const os of OSes) {
-            newPayload[os][protection].mode = ProtectionModes.prevent;
-            if (isPlatinumPlus) {
-              newPayload[os].popup[protection].enabled = event.target.checked;
-            }
-          }
-        }
-        dispatch({
-          type: 'userChangedPolicyConfig',
-          payload: { policyConfig: newPayload },
-        });
-      }
-    },
-    [dispatch, policyDetailsConfig, isPlatinumPlus]
-  );
-
-  const handleUserNotificationCheckbox = useCallback(
-    (event) => {
-      if (policyDetailsConfig) {
-        const newPayload = cloneDeep(policyDetailsConfig);
-        for (const os of OSes) {
-          newPayload[os].popup[protection].enabled = event.target.checked;
-        }
-        dispatch({
-          type: 'userChangedPolicyConfig',
-          payload: { policyConfig: newPayload },
-        });
-      }
-    },
-    [policyDetailsConfig, dispatch]
-  );
-
-  const handleCustomUserNotification = useCallback(
-    (event) => {
-      if (policyDetailsConfig) {
-        const newPayload = cloneDeep(policyDetailsConfig);
-        for (const os of OSes) {
-          newPayload[os].popup[protection].message = event.target.value;
-        }
-        dispatch({
-          type: 'userChangedPolicyConfig',
-          payload: { policyConfig: newPayload },
-        });
-      }
-    },
-    [policyDetailsConfig, dispatch]
-  );
-
-  const radioButtons = useMemo(() => {
-    return (
-      <>
-        <ConfigFormHeading>
-          <FormattedMessage
-            id="xpack.securitySolution.endpoint.policyDetailsConfig.protectionLevel"
-            defaultMessage="Protection Level"
-          />
-        </ConfigFormHeading>
-        <EuiSpacer size="xs" />
-        <RadioFlexGroup>
-          <EuiFlexItem className="no-right-margin-radio" grow={2}>
-            <ProtectionRadio
-              protectionMode={radios[0].id}
-              key={radios[0].protection + radios[0].id}
-              label={radios[0].label}
-            />
-          </EuiFlexItem>
-          <EuiFlexItem className="no-horizontal-margin-radio" grow={5}>
-            <ProtectionRadio
-              protectionMode={radios[1].id}
-              key={radios[1].protection + radios[1].id}
-              label={radios[1].label}
-            />
-          </EuiFlexItem>
-        </RadioFlexGroup>
-        {isPlatinumPlus && (
-          <>
-            <EuiSpacer size="m" />
-            <ConfigFormHeading>
-              <FormattedMessage
-                id="xpack.securitySolution.endpoint.policyDetailsConfig.userNotification"
-                defaultMessage="User Notification"
-              />
-            </ConfigFormHeading>
-            <SupportedVersionNotice optionName="malware" />
-            <EuiSpacer size="s" />
-            <EuiCheckbox
-              data-test-subj="malwareUserNotificationCheckbox"
-              id="xpack.securitySolution.endpoint.policyDetail.malware.userNotification"
-              onChange={handleUserNotificationCheckbox}
-              checked={userNotificationSelected}
-              disabled={selected === ProtectionModes.off}
-              label={i18n.translate(
-                'xpack.securitySolution.endpoint.policyDetail.malware.notifyUser',
-                {
-                  defaultMessage: 'Notify User',
-                }
-              )}
-            />
-          </>
-        )}
-        {isPlatinumPlus && userNotificationSelected && (
-          <>
-            <EuiSpacer size="s" />
-            <EuiFlexGroup gutterSize="s">
-              <EuiFlexItem grow={false}>
-                <EuiText size="xs">
-                  <h4>
-                    <FormattedMessage
-                      id="xpack.securitySolution.endpoint.policyDetailsConfig.customizeUserNotification"
-                      defaultMessage="Customize notification message"
-                    />
-                  </h4>
-                </EuiText>
-              </EuiFlexItem>
-              <EuiFlexItem grow={false}>
-                <EuiIconTip
-                  position="right"
-                  data-test-subj="malwareTooltip"
-                  content={
-                    <>
-                      <FormattedMessage
-                        id="xpack.securitySolution.endpoint.policyDetailsConfig.malware.notifyUserTooltip.a"
-                        defaultMessage="Selecting the user notification option will display a notification to the host user when malware is prevented or detected."
-                      />
-                      <EuiSpacer size="m" />
-                      <FormattedMessage
-                        id="xpack.securitySolution.endpoint.policyDetailsConfig.malware.notifyUserTooltip.b"
-                        defaultMessage="
-                    The user notification can be customized in the text box below. Bracketed tags can be used to dynamically populate the applicable action (such as prevented or detected) and the filename."
-                      />
-                    </>
-                  }
-                />
-              </EuiFlexItem>
-            </EuiFlexGroup>
-            <EuiSpacer size="xs" />
-            <EuiTextArea
-              placeholder={i18n.translate(
-                'xpack.securitySolution.endpoint.policyDetails.malware.userNotification.placeholder',
-                {
-                  defaultMessage: 'Input your custom notification message',
-                }
-              )}
-              value={userNotificationMessage}
-              onChange={handleCustomUserNotification}
-              fullWidth={true}
-              data-test-subj="malwareUserNotificationCustomMessage"
-            />
-          </>
-        )}
-      </>
-    );
-  }, [
-    radios,
-    selected,
-    isPlatinumPlus,
-    handleUserNotificationCheckbox,
-    userNotificationSelected,
-    userNotificationMessage,
-    handleCustomUserNotification,
-  ]);
-
-  const protectionSwitch = useMemo(() => {
-    return (
-      <EuiSwitch
-        label={i18n.translate(
-          'xpack.securitySolution.endpoint.policy.details.malwareProtectionsEnabled',
-          {
-            defaultMessage: 'Malware protections {mode, select, true {enabled} false {disabled}}',
-            values: {
-              mode: selected !== ProtectionModes.off,
-            },
-          }
-        )}
-        checked={selected !== ProtectionModes.off}
-        onChange={handleSwitchChange}
-      />
-    );
-  }, [handleSwitchChange, selected]);
 
   return (
     <ConfigForm
@@ -339,9 +35,10 @@ export const MalwareProtections = React.memo(() => {
       })}
       supportedOss={[OperatingSystem.WINDOWS, OperatingSystem.MAC]}
       dataTestSubj="malwareProtectionsForm"
-      rightCorner={protectionSwitch}
+      rightCorner={<ProtectionSwitch protection={protection} osList={OSes} />}
     >
-      {radioButtons}
+      <RadioButtons protection={protection} osList={OSes} />
+      {isPlatinumPlus && <UserNotification protection={protection} osList={OSes} />}
       <EuiSpacer size="m" />
       <EuiCallOut iconType="iInCircle">
         <FormattedMessage
