+++ x-pack/plugins/translations/translations/ja-JP.json
@@ -19610,12 +19610,10 @@
     "xpack.securitySolution.endpoint.policy.details.eventCollectionsEnabled": "{selected} / {total} 件のイベント収集が有効です",
     "xpack.securitySolution.endpoint.policy.details.lockedCard": "ランサムウェア保護をオンにするには、ライセンスをプラチナに更新するか、30日間の無料トライアルを開始するか、AWS、GCP、またはAzureで{cloudDeploymentLink}にサインアップしてください。",
     "xpack.securitySolution.endpoint.policy.details.malware": "マルウェア",
-    "xpack.securitySolution.endpoint.policy.details.malwareProtectionsEnabled": "マルウェア保護{mode, select, true {有効} false {無効}}",
     "xpack.securitySolution.endpoint.policy.details.platinum": "プラチナ",
     "xpack.securitySolution.endpoint.policy.details.prevent": "防御",
     "xpack.securitySolution.endpoint.policy.details.protections": "保護",
     "xpack.securitySolution.endpoint.policy.details.ransomware": "ランサムウェア",
-    "xpack.securitySolution.endpoint.policy.details.ransomwareProtectionsEnabled": "ランサムウェア保護{mode, select, true {有効} false {無効}}",
     "xpack.securitySolution.endpoint.policy.details.save": "保存",
     "xpack.securitySolution.endpoint.policy.details.settings": "設定",
     "xpack.securitySolution.endpoint.policy.details.updateConfirm.cancelButtonTitle": "キャンセル",
@@ -19627,15 +19625,11 @@
     "xpack.securitySolution.endpoint.policy.details.updateSuccessMessage": "統合{name}が更新されました。",
     "xpack.securitySolution.endpoint.policy.details.updateSuccessTitle": "成功！",
     "xpack.securitySolution.endpoint.policy.details.upgradeToPlatinum": "Elastic Platinum へのアップグレード",
-    "xpack.securitySolution.endpoint.policyDetail.malware.notifyUser": "ユーザーに通知",
-    "xpack.securitySolution.endpoint.policyDetail.ransomware.notifyUser": "ユーザーに通知",
     "xpack.securitySolution.endpoint.policyDetailOS": "オペレーティングシステム",
     "xpack.securitySolution.endpoint.policyDetails.agentsSummary.errorTitle": "エラー",
     "xpack.securitySolution.endpoint.policyDetails.agentsSummary.offlineTitle": "オフライン",
     "xpack.securitySolution.endpoint.policyDetails.agentsSummary.onlineTitle": "オンライン",
     "xpack.securitySolution.endpoint.policyDetails.agentsSummary.totalTitle": "エンドポイント",
-    "xpack.securitySolution.endpoint.policyDetails.malware.userNotification.placeholder": "カスタム通知メッセージを入力",
-    "xpack.securitySolution.endpoint.policyDetails.ransomware.userNotification.placeholder": "カスタム通知メッセージを入力",
     "xpack.securitySolution.endpoint.policyDetails.supportedVersion": "エージェントバージョン {version}",
     "xpack.securitySolution.endpoint.policyDetailsConfig.customizeUserNotification": "通知メッセージをカスタマイズ",
     "xpack.securitySolution.endpoint.policyDetailsConfig.eventingEvents": "イベント",
@@ -19645,11 +19639,7 @@
     "xpack.securitySolution.endpoint.policyDetailsConfig.mac.events.file": "ファイル",
     "xpack.securitySolution.endpoint.policyDetailsConfig.mac.events.network": "ネットワーク",
     "xpack.securitySolution.endpoint.policyDetailsConfig.mac.events.process": "プロセス",
-    "xpack.securitySolution.endpoint.policyDetailsConfig.malware.notifyUserTooltip.a": "ユーザー通知オプションを選択すると、マルウェアが防御または検出されたときに、ホストユーザーに通知を表示します。",
-    "xpack.securitySolution.endpoint.policyDetailsConfig.malware.notifyUserTooltip.b": " ユーザー通知は、以下のテキストボックスでカスタマイズできます。括弧内のタグを使用すると、該当するアクション (防御または検出など) とファイル名を動的に入力できます。",
     "xpack.securitySolution.endpoint.policyDetailsConfig.protectionLevel": "保護レベル",
-    "xpack.securitySolution.endpoint.policyDetailsConfig.ransomware.notifyUserTooltip.a": "ユーザー通知オプションを選択すると、ランサムウェアが防御または検出されたときに、ホストユーザーに通知を表示します。",
-    "xpack.securitySolution.endpoint.policyDetailsConfig.ransomware.notifyUserTooltip.b": " ユーザー通知は、以下のテキストボックスでカスタマイズできます。括弧内のタグを使用すると、該当するアクション (防御または検出など) とファイル名を動的に入力できます。",
     "xpack.securitySolution.endpoint.policyDetailsConfig.userNotification": "ユーザー通知",
     "xpack.securitySolution.endpoint.policyDetailsConfig.windows.events.dllDriverLoad": "DLL とドライバーの読み込み",
     "xpack.securitySolution.endpoint.policyDetailsConfig.windows.events.dns": "DNS",
