+++ x-pack/plugins/translations/translations/zh-CN.json
@@ -19900,12 +19900,10 @@
     "xpack.securitySolution.endpoint.policy.details.eventCollectionsEnabled": "{selected} / {total} 个事件收集已启用",
     "xpack.securitySolution.endpoint.policy.details.lockedCard": "要打开勒索软件防护，必须将您的许可证升级到白金级、开始 30 天免费试用或在 AWS、GCP 或 Azure 中实施{cloudDeploymentLink}。",
     "xpack.securitySolution.endpoint.policy.details.malware": "恶意软件",
-    "xpack.securitySolution.endpoint.policy.details.malwareProtectionsEnabled": "恶意软件防护{mode, select, true {已启用} false {已禁用}}",
     "xpack.securitySolution.endpoint.policy.details.platinum": "白金级",
     "xpack.securitySolution.endpoint.policy.details.prevent": "防御",
     "xpack.securitySolution.endpoint.policy.details.protections": "防护",
     "xpack.securitySolution.endpoint.policy.details.ransomware": "勒索软件",
-    "xpack.securitySolution.endpoint.policy.details.ransomwareProtectionsEnabled": "勒索软件防护{mode, select, true {已启用} false {已禁用}}",
     "xpack.securitySolution.endpoint.policy.details.save": "保存",
     "xpack.securitySolution.endpoint.policy.details.settings": "设置",
     "xpack.securitySolution.endpoint.policy.details.updateConfirm.cancelButtonTitle": "取消",
@@ -19918,15 +19916,11 @@
     "xpack.securitySolution.endpoint.policy.details.updateSuccessMessage": "集成 {name} 已更新。",
     "xpack.securitySolution.endpoint.policy.details.updateSuccessTitle": "成功！",
     "xpack.securitySolution.endpoint.policy.details.upgradeToPlatinum": "升级到 Elastic 白金级",
-    "xpack.securitySolution.endpoint.policyDetail.malware.notifyUser": "通知用户",
-    "xpack.securitySolution.endpoint.policyDetail.ransomware.notifyUser": "通知用户",
     "xpack.securitySolution.endpoint.policyDetailOS": "操作系统",
     "xpack.securitySolution.endpoint.policyDetails.agentsSummary.errorTitle": "错误",
     "xpack.securitySolution.endpoint.policyDetails.agentsSummary.offlineTitle": "脱机",
     "xpack.securitySolution.endpoint.policyDetails.agentsSummary.onlineTitle": "联机",
     "xpack.securitySolution.endpoint.policyDetails.agentsSummary.totalTitle": "终端",
-    "xpack.securitySolution.endpoint.policyDetails.malware.userNotification.placeholder": "输入您的定制通知消息",
-    "xpack.securitySolution.endpoint.policyDetails.ransomware.userNotification.placeholder": "输入您的定制通知消息",
     "xpack.securitySolution.endpoint.policyDetails.supportedVersion": "代理版本 {version}",
     "xpack.securitySolution.endpoint.policyDetailsConfig.customizeUserNotification": "定制通知消息",
     "xpack.securitySolution.endpoint.policyDetailsConfig.eventingEvents": "事件",
@@ -19936,11 +19930,7 @@
     "xpack.securitySolution.endpoint.policyDetailsConfig.mac.events.file": "文件",
     "xpack.securitySolution.endpoint.policyDetailsConfig.mac.events.network": "网络",
     "xpack.securitySolution.endpoint.policyDetailsConfig.mac.events.process": "进程",
-    "xpack.securitySolution.endpoint.policyDetailsConfig.malware.notifyUserTooltip.a": "选择用户通知选项后，在阻止或检测到恶意软件时将向主机用户显示通知。",
-    "xpack.securitySolution.endpoint.policyDetailsConfig.malware.notifyUserTooltip.b": " 可在下方文本框中定制用户通知。括号中的标签可用于动态填充适用操作 (如已阻止或已检测) 和文件名。",
     "xpack.securitySolution.endpoint.policyDetailsConfig.protectionLevel": "防护级别",
-    "xpack.securitySolution.endpoint.policyDetailsConfig.ransomware.notifyUserTooltip.a": "选择用户通知选项后，在阻止或检测到勒索软件时将向主机用户显示通知。",
-    "xpack.securitySolution.endpoint.policyDetailsConfig.ransomware.notifyUserTooltip.b": " 可在下方文本框中定制用户通知。括号中的标签可用于动态填充适用操作 (如已阻止或已检测) 和文件名。",
     "xpack.securitySolution.endpoint.policyDetailsConfig.userNotification": "用户通知",
     "xpack.securitySolution.endpoint.policyDetailsConfig.windows.events.dllDriverLoad": "DLL 和驱动程序加载",
     "xpack.securitySolution.endpoint.policyDetailsConfig.windows.events.dns": "DNS",
